+++ __tests__/context.js
@@ -10,13 +10,11 @@ it( 'append directives', () => {
     ctx.appendDirectives( {
         second: 2
     } );
-    expect( ctx.directives ).toEqual( {
-        first: 1,
-        second: 2
-    } );
-    expect( root.directives ).toEqual( {
-        first: 1
-    } );
+    expect( ctx.directives.first ).toEqual( 1 );
+    expect( ctx.directives.second ).toEqual( 2 );
+
+    expect( root.directives.first ).toEqual( 1 );
+    expect( root.directives.second ).toEqual( undefined );
 } );
 
 it( 'append filters', () => {
@@ -29,12 +27,71 @@ it( 'append filters', () => {
     ctx.appendFilters( {
         second: 2
     } );
-    expect( ctx.filters ).toEqual( {
-        first: 1,
-        second: 2
-    } );
-    expect( root.filters ).toEqual( {
-        first: 1
-    } );
+    expect( ctx.filters.first ).toEqual( 1 );
+    expect( ctx.filters.second ).toEqual( 2 );
+
+    expect( root.filters.first ).toEqual( 1 );
+    expect( root.filters.second ).toEqual( undefined );
 } );
 
+it( 'append filters 2 level', () => {
+    const root = createRootContext( {
+        filters: {
+            first: 1
+        }
+    } );
+    const ctx = createChildContext( { ctx: root } );
+    ctx.appendFilters( {
+        second: 2
+    } );
+    const nestedCtx = createChildContext( { ctx } );
+    expect( nestedCtx.filters.first ).toEqual( 1 );
+    expect( nestedCtx.filters.second ).toEqual( 2 );
+
+    expect( ctx.filters.first ).toEqual( 1 );
+    expect( ctx.filters.second ).toEqual( 2 );
+
+    expect( root.filters.first ).toEqual( 1 );
+    expect( root.filters.second ).toEqual( undefined );
+} );
+
+
+it( 'append filters to parent after create child', () => {
+    const root = createRootContext( {
+        filters: {
+            first: 1
+        }
+    } );
+    const ctx = createChildContext( { ctx: root } );
+    root.appendFilters( {
+        second: 2
+    } );
+
+    expect( ctx.filters.first ).toEqual( 1 );
+    expect( ctx.filters.second ).toEqual( undefined );
+
+    expect( root.filters.first ).toEqual( 1 );
+    expect( root.filters.second ).toEqual( 2 );
+} );
+
+it( 'append filters to parent after create child2 level', () => {
+    const root = createRootContext( {
+        filters: {
+            first: 1
+        }
+    } );
+    const ctx = createChildContext( { ctx: root } );
+    const nestedCtx = createChildContext( { ctx } );
+    root.appendFilters( {
+        second: 2
+    } );
+
+    expect( nestedCtx.filters.first ).toEqual( 1 );
+    expect( nestedCtx.filters.second ).toEqual( undefined );
+
+    expect( ctx.filters.first ).toEqual( 1 );
+    expect( ctx.filters.second ).toEqual( undefined );
+
+    expect( root.filters.first ).toEqual( 1 );
+    expect( root.filters.second ).toEqual( 2 );
+} );